requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
croniter==6.0.0
cryptography==44.0.0
decorator==5.1.1
executing==2.1.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-cron==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
